backend/requirements.txt
flask
pymonogo
coinbase
python-dotenv